backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.125.0
h11==0.16.0
idna==3.11
pymongo==4.15.5
python-dotenv==1.2.1
python-multipart==0.0.21
rsa==4.9.1
six==1.17.0
starlette==0.50.0
supabase==2.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
